Modify nav menu & MyProfileLayout Component
